requirements.txt
streamlit>=1.38.0
pandas>=2.2.0
openpyxl>=3.1.2
plotly>=5.22.0
altair>=5.3.0
python-dotenv>=1.0.1
google-generativeai>=0.7.0
